refactor: split Rich Text Input file into smaller files (#4231)
